feat: added notification component to toolbar
