+++ src/contentscript/overlay/root/components/OverlayToolbar/OverlayToolbar.module.scss
@@ -528,6 +528,7 @@
 @keyframes animate-notification {
   to {
     transform: translateX(400px);
+    opacity: 0;
   }
 }
